Renamespace api and frontend token create/verify tests
